micro_services/counting/requirements.txt
python-dotenv==0.14.0
confluent-kafka==1.5.0
SQLAlchemy==1.3.13
mysqlclient==2.0.1
